Fix int64 in Gather and TopK

+++ runtime/onnion_runtime/gather.py
@@ -7,4 +7,4 @@ class Gather:
         self.axis = kwargs.get("axis", 0)
 
     def run(self, data, indices):
-        return [np.take(data, indices, axis=self.axis)]
+        return [np.take(data, indices.astype(int), axis=self.axis)]
